Improvement: Relocated BodyLocation Enum into Medical Package
